Fixed FileChannel not being closed properly

+++ src/main/java/com/mndk/kvm2m/core/util/file/DirectoryManager.java
@@ -5,11 +5,19 @@ import java.io.File;
 public class DirectoryManager {
 
 	public static void deleteDirectory(File file) {
-		for(File subFile : file.listFiles()) {
+		File[] files = file.listFiles();
+		if(files == null) return;
+		for(File subFile : files) {
 			if(subFile.isDirectory()) deleteDirectory(subFile);
-			else subFile.delete();
+			else {
+				if(!subFile.delete()) {
+					throw new RuntimeException("Failed to delete the file");
+				}
+			}
 		}
-		file.delete();
+		if(!file.delete()) {
+			throw new RuntimeException("Failed to delete the directory");
+		}
 	}
 
 }
